updating with renamed modules and added to the license
